Switch to new PricingManager
- remove PricingManager
- rename PricingManagerV2 to PricingManager
- remove CatalogManager
