--- src/Foundation/Connect/code/Managers/Contracts/ICatalogManager.cs
@@ -1,34 +1,0 @@
-//    Copyright 2020 EPAM Systems, Inc.
-// 
-//    Licensed under the Apache License, Version 2.0 (the "License");
-//    you may not use this file except in compliance with the License.
-//    You may obtain a copy of the License at
-// 
-//      http://www.apache.org/licenses/LICENSE-2.0
-// 
-//    Unless required by applicable law or agreed to in writing, software
-//    distributed under the License is distributed on an "AS IS" BASIS,
-//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//    See the License for the specific language governing permissions and
-//    limitations under the License.
-
-namespace Wooli.Foundation.Connect.Managers
-{
-    using System.Collections.Generic;
-
-    using Models.Catalog;
-
-    using Sitecore.Commerce.Services.Prices;
-    using Sitecore.Data.Items;
-
-    public interface ICatalogManager
-    {
-        ManagerResponse<GetProductBulkPricesResult, bool> GetProductBulkPrices(List<Product> products);
-
-        void GetProductPrice(Product product);
-
-        decimal? GetProductRating(Item productItem);
-
-        void GetStockInfo(Product product, string shopName);
-    }
-}
