+++ src/Foundation/Connect/code/Managers/Pricing/PricingManager.cs
@@ -1,4 +1,4 @@
-//    Copyright 2020 EPAM Systems, Inc.
+﻿//    Copyright 2020 EPAM Systems, Inc.
 // 
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -12,71 +12,63 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-namespace Wooli.Foundation.Connect.Managers
+namespace Wooli.Foundation.Connect.Managers.Pricing
 {
     using System;
     using System.Collections.Generic;
+
+    using Base.Models.Logging;
+    using Base.Services.Logging;
 
     using DependencyInjection;
 
     using Providers.Contracts;
 
-    using Sitecore.Commerce.Entities.Prices;
     using Sitecore.Commerce.Services.Prices;
     using Sitecore.Diagnostics;
 
     using GetProductBulkPricesRequest = Sitecore.Commerce.Engine.Connect.Services.Prices.GetProductBulkPricesRequest;
     using GetProductPricesRequest = Sitecore.Commerce.Engine.Connect.Services.Prices.GetProductPricesRequest;
 
-    [Service(typeof(IPricingManager))]
-    public class PricingManager : IPricingManager
+    [Service(typeof(IPricingManager), Lifetime = Lifetime.Singleton)]
+    public class PricingManager : BaseManager, IPricingManager
     {
-        // TODO: Move this array to CatalogManager
-        private static readonly string[] DefaultPriceTypeIds = new string[5]
-        {
-            "List", "Adjusted", "LowestPricedVariant", "LowestPricedVariantListPrice", "HighestPricedVariant"
-        };
-
         private readonly PricingServiceProvider pricingServiceProvider;
 
-        public PricingManager(IConnectServiceProvider connectServiceProvider)
+        public PricingManager(IConnectServiceProvider connectServiceProvider, ILogService<CommonLog> logService)
+            : base(logService)
         {
             Assert.ArgumentNotNull(connectServiceProvider, nameof(connectServiceProvider));
+
             this.pricingServiceProvider = connectServiceProvider.GetPricingServiceProvider();
         }
 
-        public ManagerResponse<GetProductBulkPricesResult, IDictionary<string, Price>> GetProductBulkPrices(
+        public GetProductBulkPricesResult GetProductBulkPrices(
             string catalogName,
             IEnumerable<string> productIds,
             params string[] priceTypeIds)
         {
-            if (priceTypeIds == null)
-            {
-                priceTypeIds = DefaultPriceTypeIds;
-            }
+            Assert.ArgumentNotNullOrEmpty(catalogName, nameof(catalogName));
+            Assert.ArgumentNotNull(productIds, nameof(productIds));
+            Assert.ArgumentNotNull(priceTypeIds, nameof(priceTypeIds));
 
             var bulkPricesRequest = new GetProductBulkPricesRequest(catalogName, productIds, priceTypeIds)
             {
                 DateTime = DateTime.UtcNow
             };
 
-            var productBulkPrices = this.pricingServiceProvider.GetProductBulkPrices(bulkPricesRequest);
-            var result = productBulkPrices.Prices ?? new Dictionary<string, Price>();
-            return new ManagerResponse<GetProductBulkPricesResult, IDictionary<string, Price>>(
-                productBulkPrices,
-                result);
+            return this.Execute(bulkPricesRequest, this.pricingServiceProvider.GetProductBulkPrices);
         }
 
-        public ManagerResponse<GetProductPricesResult, IDictionary<string, Price>> GetProductPrices(
+        public GetProductPricesResult GetProductPrices(
             string catalogName,
             string productId,
             bool includeVariants,
             params string[] priceTypeIds)
         {
-            if (priceTypeIds == null)
-            {
-                priceTypeIds = DefaultPriceTypeIds;
-            }
+            Assert.ArgumentNotNullOrEmpty(catalogName, nameof(catalogName));
+            Assert.ArgumentNotNullOrEmpty(productId, nameof(productId));
+            Assert.ArgumentNotNull(priceTypeIds, nameof(priceTypeIds));
 
             var pricesRequest = new GetProductPricesRequest(catalogName, productId, priceTypeIds)
             {
@@ -84,11 +76,7 @@ namespace Wooli.Foundation.Connect.Managers
                 IncludeVariantPrices = includeVariants
             };
 
-            var serviceProviderResult = this.pricingServiceProvider.GetProductPrices(pricesRequest);
-
-            return new ManagerResponse<GetProductPricesResult, IDictionary<string, Price>>(
-                serviceProviderResult,
-                serviceProviderResult.Prices ?? new Dictionary<string, Price>());
+            return this.Execute(pricesRequest, this.pricingServiceProvider.GetProductPrices);
         }
     }
 }
