+++ src/Foundation/Connect/code/Wooli.Foundation.Connect.csproj
@@ -600,8 +600,8 @@
     <Compile Include="Models\Catalog\StockStatus.cs" />
     <Compile Include="Models\Payment\PaymentClientTokenResult.cs" />
     <Compile Include="Services\Search\ISearchService.cs" />
-    <Compile Include="Managers\Pricing\IPricingManagerV2.cs" />
-    <Compile Include="Managers\Pricing\PricingManagerV2.cs" />
+    <Compile Include="Managers\Pricing\IPricingManager.cs" />
+    <Compile Include="Managers\Pricing\PricingManager.cs" />
     <Compile Include="Managers\Order\IOrderManagerV2.cs" />
     <Compile Include="Managers\PaymentManager.cs" />
     <Compile Include="Services\Search\SearchService.cs" />
@@ -627,18 +627,14 @@
     <Compile Include="Models\Catalog\Variant.cs" />
     <Compile Include="Managers\AccountManager.cs" />
     <Compile Include="Managers\CartManager.cs" />
-    <Compile Include="Managers\CatalogManager.cs" />
     <Compile Include="Managers\Contracts\ICartManager.cs" />
     <Compile Include="Managers\Contracts\IInventoryManager.cs" />
     <Compile Include="Managers\Contracts\IOrderManager.cs" />
-    <Compile Include="Managers\Contracts\IPricingManager.cs" />
     <Compile Include="Managers\Contracts\IAccountManager.cs" />
-    <Compile Include="Managers\Contracts\ICatalogManager.cs" />
     <Compile Include="Managers\Contracts\IShippingManager.cs" />
     <Compile Include="Managers\InventoryManager.cs" />
     <Compile Include="Managers\ManagerResponse.cs" />
     <Compile Include="Managers\OrderManager.cs" />
-    <Compile Include="Managers\PricingManager.cs" />
     <Compile Include="Managers\ShippingManager.cs" />
     <Compile Include="Models\CartLineArgument.cs" />
     <Compile Include="Models\Search\SearchResults.cs" />
