+++ src/Foundation/Connect/tests/Wooli.Foundation.Connect.Tests.csproj
@@ -571,7 +571,7 @@
     <Compile Include="Mappers\Profiles\AddressProfileTests.cs" />
     <Compile Include="Mappers\Profiles\CatalogProfileTests.cs" />
     <Compile Include="Services\Search\SearchServiceTests.cs" />
-    <Compile Include="Managers\Pricing\PricingManagerV2Tests.cs" />
+    <Compile Include="Managers\Pricing\PricingManagerTests.cs" />
     <Compile Include="Managers\Shipping\ShippingManagerV2Tests.cs" />
     <Compile Include="Mappers\Profiles\SearchProfileTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
